Commit by RM.
Added sellers number of successful transactions to yield.

Fixed a URL-parameter issue which took up space in the log when the user could not be identified. Manual inspection showed some items did not have an associated vendor. A try block issues a warning when this happens, and the vendor username is set as ''.

Another issue was the URL-parameter 'lid'. This only occurs once, seemingly on the frontpage, so a warning is issued and the error is handled.

Item-prices yielded empty fields, which was likely caused by updated layout on the website. A new regular expression was added which solves this and returns the price in BTC.

Multilisting option were not collected, but the addition of an @ to the xpath solved it.

Usernames were collected incorrectly including the rating of the user and the number of purchases. We now only collect the username.

The code for extracting ratings was only finding 5/5's. This has been fixed and we now catch different ratings.