A minor fix. Lately some problematic usernames have been observed, which resembled concatenated ussernames. Upon inspection it was found that due to some weird nested structure of threads, we were not properly collecting usernames. The issue has been resolved and usernames are now correctly collected. However, note the following:
The problematic username GoodFellasuddenwoebegoneunknown13sedeactxanaxnationnzhappydaze previously had both a post and a user id associated. After fixing the scraping of the username, a new user is created in the DB which is now attached to the post. The old user id remains though. During later validation, it might seem as if we are encountering this username again because it has not been purged from the DB.